REFACT: Tests correction + refact JobService to handle errors

+++ api/src/main/java/com/jobstream/api/service/JobService.java
@@ -1,16 +1,16 @@
 package com.jobstream.api.service;
 
+import com.jobstream.api.exception.ResourceConflictException;
+import com.jobstream.api.exception.ResourceNotFoundException;
 import com.jobstream.api.mapper.JobMapper;
 import com.jobstream.api.repository.JobRepository;
 import com.jobstream.dto.JobDto;
 import com.jobstream.dto.JobRequestDto;
-import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -22,7 +22,7 @@ public class JobService {
     public JobDto getJobById(Long id){
         return jobRepository.findById(id)
                 .map(jobMapper::toDto)
-                .orElseThrow(() -> new EntityNotFoundException("Job non trouvé avec l'id : " + id));
+                .orElseThrow(() -> new ResourceNotFoundException("Job non trouvé avec l'id : " + id));
     }
 
     public Page<JobDto> getJobs(Pageable pageable) {
@@ -31,12 +31,18 @@ public class JobService {
     }
 
     @Transactional
-    public JobDto saveJob(JobRequestDto jobDto){
-        return jobMapper.toDto(jobRepository.save(jobMapper.toEntity(jobDto)));
+    public JobDto saveJob(JobRequestDto jobRequestDto){
+        if (jobRepository.existsByExternalId(jobRequestDto.getExternalId())) {
+            throw new ResourceConflictException("Job déjà sauvegardé avec l'id externe : " + jobRequestDto.getExternalId());
+        }
+        return jobMapper.toDto(jobRepository.save(jobMapper.toEntity(jobRequestDto)));
     }
 
     @Transactional
     public void deleteJob(Long id){
+        if (!jobRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Job non trouvé avec l'id : " + id);
+        }
         jobRepository.deleteById(id);
     }
 }
